Completed relocation

+++ src/Common/Symbol.hpp
@@ -7,13 +7,13 @@ enum SymbolType : uint8_t { SYMBOL_TEXT, SYMBOL_DATA, SYMBOL_NONE };
 struct Symbol {
   std::string name;
   SymbolType type;
-  uint16_t address;
+  uint32_t address;
   bool isGlobal;
 };
 
 struct SymbolForMap {
   SymbolType type;
-  uint16_t address;
+  uint32_t address;
 };
 
 bool isValidSymbolName(std::string &symbolName);
